Fixed some bugs in length dataset

+++ executors/trainers/bp_trainer.py
@@ -78,7 +78,7 @@ class BackpropagationTrainer(Trainer):
     def train_on_batch(self, index, sampled_batch, print_info):
         # Load data
         image_tensor = sampled_batch['image'].to(self.device)
-        labels = sampled_batch['struct_info'].to(self.device)
+        labels = sampled_batch['struct_info'].to(self.device).float()
 
         # zero the parameter gradients
         self.optimizer.zero_grad()
